requirements.txt
fastapi
uvicorn
python-dotenv
supabase
google-generativeai
pydantic
requests
pillow
youtube-transcript-api 